requirements.txt
fastapi
uvicorn
gpiozero
picamera
